LayoutTests/fast/parser/fast-html-parser-consume-entity.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
</head>
<body>
<script src="../../resources/js-test.js"></script>
<div id="testDiv"</div>
<script>
description("Tests for the HTML entity parsing by our HTML fast parser.");

let div = document.getElementById("testDiv");
div.innerHTML = "Genius Nicer Dicer Plus | 18&nbsp&hellip;";
shouldBeEqualToString("div.innerHTML", "Genius Nicer Dicer Plus | 18&nbsp;…");
div.innerHTML = "&nbsp&a";
shouldBeEqualToString("div.innerHTML", "&nbsp;&amp;a");
div.innerHTML = "&nbsp&";
shouldBeEqualToString("div.innerHTML", "&nbsp;&amp;");
div.innerHTML = "&nbsp-";
shouldBeEqualToString("div.innerHTML", "&nbsp;-");
</script>
</body>
</html>
